refactor: moved blocking buffer in utils
