+++ internal/utils/buffer.go
@@ -1,4 +1,4 @@
-package steamcm
+package utils
 
 import (
 	"bytes"
@@ -6,12 +6,13 @@ import (
 )
 
 type buffer struct {
-	buff bytes.Buffer
-	mu   sync.Mutex
-	cond *sync.Cond
+	buff   bytes.Buffer
+	mu     sync.Mutex
+	cond   *sync.Cond
+	closed bool
 }
 
-func NewBuffer() *buffer {
+func NewBlockingBuffer() *buffer {
 	buff := &buffer{}
 	buff.cond = sync.NewCond(&buff.mu)
 	return buff
@@ -20,8 +21,7 @@ func NewBuffer() *buffer {
 func (b *buffer) Read(p []byte) (int, error) {
 	b.mu.Lock()
 	defer b.mu.Unlock()
-
-	if b.buff.Len() == 0 {
+	if b.buff.Len() == 0 && !b.closed {
 		b.cond.Wait()
 	}
 	return b.buff.Read(p)
@@ -34,4 +34,12 @@ func (b *buffer) Write(p []byte) (int, error) {
 	n, err := b.buff.Write(p)
 	b.cond.Signal()
 	return n, err
+}
+
+func (b *buffer) Close() error {
+	if !b.closed {
+		b.closed = true
+		b.cond.Broadcast()
+	}
+	return nil
 }
